improve: reduce copy paste in tests
