+++ algebird-test/src/main/scala-2.12+/com/twitter/algebird/BasePropertiesCompat.scala
@@ -1,0 +1,9 @@
+package com.twitter.algebird
+
+private[algebird] trait BasePropertiesCompat {
+  def isNonZero[V: Semigroup](v: V): Boolean =
+    implicitly[Semigroup[V]] match {
+      case mon: Monoid[?] => mon.isNonZero(v)
+      case _              => true
+    }
+}
